requirements.txt
fastapi
uvicorn
streamlit
google-genai
python-dotenv
pylint
radon
flake8